Generate a matrix of two-phase relative permeability data for the water-oil system using the modified Brooks-Corey model

The 'kr2p_ow()' creates a table of two-phase water and oil relative permeability data for water and oil saturation values between zero and one.

kr2p_ow(SWCON, SWCRIT, SOIRW, SORW, KRWIRO, KROCW, NW, NOW, NP)

Arguments

SWCON
connate water saturation, fraction

SWCRIT
critical water saturation, fraction

SOIRW
irreducible oil saturation, fraction

SORW
residual oil saturation, fraction

KRWIRO
water relative permeability at irreducible oil

KROCW
oil relative permeability at connate water

NW
exponent term for calculating krw

NOW
exponent term for calculating krow

NP
number of saturation points in the table, the maximum acceptable value is 501

Value

A matrix with water saturation, oil saturation, water relative permeability, and oil relative permeability values, respectively

References

Brooks RH, Corey AT (1964). “Hydraulic Properties of Porous Media.” Hydrology Papers, 27. https://mountainscholar.org/bitstream/handle/10217/61288/HydrologyPapers_n3.pdf.

Examples

rel_perm_wo <- kr2p_ow(0.15, 0.2, 0.15, 0.15, 0.4, 1, 3, 2, 101)